requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.2
coloredlogs==15.0.1
dataclasses-json==0.5.7
decorator==5.1.1
duckdb==0.8.0
exceptiongroup==1.1.1
fastapi==0.96.1
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jq==1.4.1
jsonschema==4.17.3
langchain==0.0.196
langchainplus-sdk==0.0.8
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
posthog==3.0.1
protobuf==4.23.2
pulsar-client==3.2.0
pyarrow==12.0.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypdf==3.9.1
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
